OSSAMA/requirements.txt
pyusb
adb-shell
tk
sdkmanager
